pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"


[tool.poetry]
name = "focus"
description = "Focus"
version = "0.1.0"
readme = "README.md"
authors = ["O.Dev"]


[tool.poetry.dependencies]
python = "3.9.9"

django = "^3.2"
django-split-settings = "^1.1"
django-axes = "^5.31"
django-csp = "^3.7"
django-health-check = "^3.16"
django-http-referrer-policy = "^1.1"
django-permissions-policy = "^4.7"
django-stubs-ext = "^0.3"

psycopg2-binary = "^2.9"
gunicorn = "^20.0"
python-decouple = "^3.5"
bcrypt = "^3.2"
structlog = "^21.5"
djangorestframework = "3.13.1"
retailcrm = "5.1.2"
pydantic = "^1.9.0"
dadata = "^21.10.1"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2"
django-querycount = "^0.7"
django-migration-linter = "^3.0"
django-extra-checks = "^0.11"
django-coverage-plugin = "^2.0"
nplusone = "^1.0"

wemake-python-styleguide = "^0.16"
flake8-pytest-style = "^1.6"
flake8-django = "^1.1"
flake8-logging-format = "^0.6"
nitpick = "^0.31"

pytest = "^6.2"
pytest-django = "^4.5"
pytest-cov = "^3.0"
pytest-randomly = "^3.11"
pytest-deadfixtures = "^2.2"
pytest-testmon = "^1.2"
pytest-timeout = "^2.1"
django-test-migrations = "^1.2"
hypothesis = "^6.36"

mypy = "^0.931"
django-stubs = "^1.9"

yamllint = "^1.26"
safety = "^1.10"
dotenv-linter = "^0.3"
polint = "^0.4"
dennis = "^0.9"
dump-env = "^1.3"
ipython = "^7.31"
